DDIT Java Study 220609

+++ src/DDitJavaClass/Exercise21.java
@@ -6,9 +6,9 @@ public class Exercise21 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("가로의 길이는?(단위:m): ");
-        int strNum1 = scanner.nextInt();
+        double strNum1 = scanner.nextDouble();
         System.out.println("세로의 길이는?(단위:m): ");
-        int strNum2 = scanner.nextInt();
+        double strNum2 = scanner.nextDouble();
 
         double width = 3.7;
         double height = 7.5;
